move the custom rtree object into a separate module
