+++ src/rtreeobject.rs
@@ -1,0 +1,35 @@
+use geo::Polygon;
+use rstar::{primitives::CachedEnvelope, RTreeObject};
+use serde::{Deserialize, Serialize};
+use std::ops::Deref;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct NumberedCell {
+    index: usize,
+    envelope: CachedEnvelope<Polygon>,
+}
+
+impl NumberedCell {
+    pub fn new(idx: usize, geom: Polygon) -> Self {
+        NumberedCell {
+            index: idx,
+            envelope: CachedEnvelope::<Polygon>::new(geom),
+        }
+    }
+
+    pub fn geometry(&self) -> &Polygon {
+        self.envelope.deref()
+    }
+
+    pub fn index(&self) -> usize {
+        self.index
+    }
+}
+
+impl RTreeObject for NumberedCell {
+    type Envelope = <CachedEnvelope<Polygon> as RTreeObject>::Envelope;
+
+    fn envelope(&self) -> Self::Envelope {
+        self.envelope.envelope()
+    }
+}
